commit to jwt-auth-03

+++ src/main/java/com/nchl/JWT/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<CreditorUser, Integer> {
+public interface UserRepository extends JpaRepository<CreditorUser, Long> {
     Optional<CreditorUser> findByEmail(String email);
     Optional<CreditorUser> findById(Integer id);
